requirements.txt
pytest-django==2.6.2
django==1.6
pytest-xdist
